pip_requirements.txt
sentence-transformers
langchain_community
openai